Change also submission.length

+++ acoustid/tables.py
@@ -159,7 +159,7 @@ submission = Table(
     Column("source_id", Integer, nullable=True),  # XXX deprecated
     # fingerprint
     Column("fingerprint", ARRAY(Integer), nullable=False),
-    Column("length", SmallInteger, CheckConstraint("length>0"), nullable=False),
+    Column("length", Integer, CheckConstraint("length>0"), nullable=False),
     Column("bitrate", SmallInteger, CheckConstraint("bitrate>0")),
     Column("format", String),
     Column("format_id", Integer),  # XXX deprecated
